hyprland: Fix block comments formatting
Now that my Gentoo machine is back to hyprland >=0.40, I no longer need
the weird formatting on the block comments. Before that version, comment
lines *had* to start with `# `, and not having the space caused the line
to not register as a comment, thus creating syntax errors when parsing
the config. This is no longer a problem in newer versions.